Cleanup timing across the engine

+++ disabled/seek_sample.cpp
@@ -31,7 +31,7 @@ public:
         return true;
     }
 
-    void update(double dt) {
+    void update(float dt) {
         set_velocity(follower_->seek(target));
 
         actor()->move_to(position() + (velocity() * dt));
@@ -72,7 +72,7 @@ public:
         );
     }
 
-    void do_step(double dt) {
+    void do_step(float dt) {
         dot_->update(dt);
     }
