+++ samples/terrain_sample.cpp
@@ -95,7 +95,7 @@ public:
         window->enable_pipeline(pipeline_id_);
     }
 
-    void fixed_update(double dt) override {
+    void fixed_update(float dt) override {
         auto stage = window->stage(stage_id_);
         stage->actor(terrain_actor_id_)->rotate_global_y_by(smlt::Degrees(dt * 5.0));
     }
