+++ simulant/interfaces/updateable.h
@@ -17,22 +17,22 @@ public:
      * Non-Virtual Interface. Simulant calls these underscore prefixed
      * functions, and subclasses implement the more nicely named ones
      */
-    virtual void _update_thunk(double dt) {
+    virtual void _update_thunk(float dt) {
         update(dt);
     }
 
-    virtual void _late_update_thunk(double dt) {
+    virtual void _late_update_thunk(float dt) {
         late_update(dt);
     }
 
-    virtual void _fixed_update_thunk(double step) {
+    virtual void _fixed_update_thunk(float step) {
         fixed_update(step);
     }
 
 private:
-    virtual void update(double dt) {}
-    virtual void late_update(double dt) {}
-    virtual void fixed_update(double step) {}
+    virtual void update(float dt) {}
+    virtual void late_update(float dt) {}
+    virtual void fixed_update(float step) {}
 };
 
 }
